// client/src/types.ts
export interface ServiceItem {
  title: string
  description: string
  icon: string
}

export interface CourseItem {
  title: string
  duration: string
  description: string
  technologies: string[]
  image: string
}

export interface TestimonialItem {
  name: string
  title: string
  quote: string
}

export interface FaqItem {
  question: string
  answer: string
}